==> 00_Requirements.ipynb <==
# <font color='Crimson'><b>Requirements</b></font>

!pip install numpy
!pip install pandas
!pip install matplotlib
!pip install seaborn
!pip install scipy
!pip install scikit-learn
!pip install tqdm
!pip install IPython
!pip install statistics
!pip install statsmodels
