==> 01_Data Preparation.ipynb <==
# <font color='Crimson'><b>DATA PREPARATION</b></font>

#Import packages:
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

#Import data set
df_dataset = pd.read_csv('00_Dataset.csv',sep=';')
df_dataset.head()

print('Number of samples:',df_dataset.shape[0])
print('Number of features:',df_dataset.shape[1])

#Update columns format:
df_dataset['SESSO'] = df_dataset['SESSO']-1
df_dataset['SESSO'] = df_dataset['SESSO'].astype('category')

df_dataset['Candidemia'] = df_dataset['Candidemia'].astype('uint8')
df_dataset['Candidemia'] = df_dataset['Candidemia'].astype('category')

df_dataset['Infezione Mista'] = df_dataset['Infezione Mista'].astype('uint8')
df_dataset['Infezione Mista'] = df_dataset['Infezione Mista'].astype('category')

df_dataset['Colonizzazione Candida 30 giorni precedenti']=((df_dataset['Colonizzazione Candida 30 giorni precedenti'])*1).astype('category')

df_dataset['Colonizzazione Candida Multifocale']=((df_dataset['Colonizzazione Candida Multifocale'])*1).astype('category')

df_dataset['Colonizzazione Candida Basse vie Respiratorie'] = df_dataset['Colonizzazione Candida Basse vie Respiratorie'].astype('category')

df_dataset['Colonizzazione Candida Urinario'] = df_dataset['Colonizzazione Candida Urinario'].astype('category')

df_dataset['Colonizzazione Candida Digestivo'] = df_dataset['Colonizzazione Candida Digestivo'].astype('category')

df_dataset.head()

#### <font color='indianred'><b>EXPLORATORY DATA ANALYSIS</b></font>

#Gender:
print('Number of patients:', len(set(df_dataset['ID_PAZIENTE'])))
genere = df_dataset[['ID_PAZIENTE', 'SESSO']]
genere_unique = genere.drop_duplicates(subset='ID_PAZIENTE')
print('Males:',genere_unique['SESSO'].value_counts()[0],'(',round(genere_unique['SESSO'].value_counts(normalize=True)[0] * 100,1),'%)')
print('Females:',genere_unique['SESSO'].value_counts()[1],'(',round(genere_unique['SESSO'].value_counts(normalize=True)[1] * 100,1),'%)')

#Age at the hospitalization:
sns.boxplot(x=df_dataset["ETA_EPISODIO"])
plt.show()
df_dataset['ETA_EPISODIO'].describe()

#Hospitalization area:
print('Number of episodes for each hospital area:\n', df_dataset['DESCR_REPARTO'].value_counts())
print('--------------------------------------------------------------')
print('Number of episodes with unknown hospital area:',df_dataset['DESCR_REPARTO'].isnull().sum())

#Reference micro-organism:
print('Number of distinct reference micro-organisms:', len(set(df_dataset['MICRORGANISMO_RIFERIMENTO'])))
print('--------------------------------------------------------------')
print('Five most frequent reference micro-organisms:\n', df_dataset['MICRORGANISMO_RIFERIMENTO'].value_counts().head())

#Episodes of Candidemia:
print('Candidemias:',df_dataset['Candidemia'].value_counts()[1],'(',round(df_dataset['Candidemia'].value_counts(normalize=True)[1] * 100,1),'%)')
print('Bacteraemias:',df_dataset['Candidemia'].value_counts()[0],'(',round(df_dataset['Candidemia'].value_counts(normalize=True)[0] * 100,1),'%)')

#Mixed infections:
print('Mixed infections (classified as Candidemias):',df_dataset['Infezione Mista'].value_counts()[1])

#Colonisation by Candida during the previous 30 days:
print('Number of episodes with a previous colonisation by Candida within 30 days:\n', df_dataset['Colonizzazione Candida 30 giorni precedenti'].value_counts())
print('--------------------------------------------------------------')
print('Number of missingness:',df_dataset['Colonizzazione Candida 30 giorni precedenti'].isnull().sum())

#Multi-focal colonisation by Candida must be 0 if the number of explored sites are less than 2:
df_dataset["Colonizzazione Candida Multifocale 2"] = df_dataset["Colonizzazione Candida Multifocale"]
for i in range(df_dataset.shape[0]):
    if (df_dataset['Numero Siti Esplorati'][i] == 0) or (df_dataset['Numero Siti Esplorati'][i] == 1):
        df_dataset['Colonizzazione Candida Multifocale 2'][i] = np.nan
df_dataset = df_dataset.iloc[:, list(range(9)) + [278] + list(range(10,278))]
print('Number of episodes with multi-focal colonisation by Candida:\n', df_dataset['Colonizzazione Candida Multifocale 2'].value_counts())
print('--------------------------------------------------------------')
print('Number of missingness:',df_dataset['Colonizzazione Candida Multifocale 2'].isnull().sum())

#Sites explored:
print('Number of sites explored per episode:\n', df_dataset['Numero Siti Esplorati'].value_counts())
print('--------------------------------------------------------------')

print('Number of colonizations by Candida in the lower respiratory tract:\n', df_dataset['Colonizzazione Candida Basse vie Respiratorie'].value_counts())
print('--------------------------------------------------------------')
print('Number of missingness:',df_dataset['Colonizzazione Candida Basse vie Respiratorie'].isnull().sum())

print('Number of colonizations by Candida in the urinary system:\n', df_dataset['Colonizzazione Candida Urinario'].value_counts())
print('--------------------------------------------------------------')
print('Number of missingness:',df_dataset['Colonizzazione Candida Urinario'].isnull().sum())

print('Number of colonizations by Candida in the digestive tract:\n', df_dataset['Colonizzazione Candida Digestivo'].value_counts())
print('--------------------------------------------------------------')
print('Number of missingness:',df_dataset['Colonizzazione Candida Digestivo'].isnull().sum())

#Assign None to laboratory tests with value 99999:
for i in range(df_dataset.shape[0]):
    for j in range(14, df_dataset.shape[1]):
        if pd.isna(df_dataset.iloc[i, j]):
            df_dataset.iloc[i, j] = None
        elif df_dataset.iloc[i, j] == 99999:
            df_dataset.iloc[i, j] = None

#### <font color='indianred'><b>EXPORT DATA SET</b></font>

#Export data set to csv file:
df_dataset.to_csv('01_Dataset.csv', sep=',')
