==> 03_Dataset split.ipynb <==
# <font color='Crimson'><b>LEARNING AND TEST SPLIT</b></font>

#Import packages:
import pandas as pd
from sklearn.model_selection import train_test_split

#Import dataset:
df_dataset = pd.read_csv('02_2_Dataset.csv',sep=',',index_col=0)
df_dataset.head()

#Features and outcome splitting:
X = df_dataset.drop(df_dataset.columns[[6]], axis=1)
y = df_dataset.iloc[:,6] 

#Training and test splitting:
X_training, X_test, y_training, y_test = train_test_split(X,y,train_size=0.7,test_size=0.3,
                                                          random_state=100,shuffle=True,stratify=y)

#The proportion of candidemias are mantained in the two sets?
print('Number of candidemias in the data set:', pd.value_counts(df_dataset['CANDIDEMIA'])[1], 
      '(',round(pd.value_counts(df_dataset['CANDIDEMIA'], normalize=True)[1]*100,2),'%) of',
      df_dataset.shape[0], 'samples')
print('--------------------------------------------------------------')
print('Number of candidemias in the TRAINING SET:', pd.value_counts(y_training)[1], 
      '(',round(pd.value_counts(y_training, normalize=True)[1]*100,2),'%) of',
      y_training.shape[0], 'samples')
print('--------------------------------------------------------------')
print('Number of candidemias in the TEST SET:', pd.value_counts(y_test)[1], 
      '(',round(pd.value_counts(y_test, normalize=True)[1]*100,2),'%) of',
      y_test.shape[0], 'samples')

#### <font color='indianred'><b>EXPORT DATA SETS</b></font>

#Export data sets to csv file:

#Training set:
X_training.to_csv('03_X_training.csv', sep=',')
y_training.to_csv('03_y_training.csv', sep=',')

#Test set:
X_test.to_csv('03_X_test.csv', sep=',')
y_test.to_csv('03_y_test.csv', sep=',')
