==> 04_Imputation.ipynb <==
# <font color='Crimson'><b>IMPUTATION</b></font>

#Import packages:
import pandas as pd
from sklearn.experimental import enable_iterative_imputer 
from sklearn.impute import IterativeImputer
from sklearn.neighbors import KNeighborsRegressor

#Import training set features:
X_training = pd.read_csv('03_X_training.csv',sep=',',index_col=0)
X_training.head()

#Import training set outcome:
y_training = pd.read_csv('03_y_training.csv',sep=',',index_col=0)
y_training.head()

#Import test set features:
X_test = pd.read_csv('03_X_test.csv',sep=',',index_col=0)
X_test.head()

#Import test set outcome:
y_test = pd.read_csv('03_y_test.csv',sep=',',index_col=0)
y_test.head()

#Merge training set features and outcome in training set:
training = X_training.drop(X_training.columns[[0,3,4,5]], axis=1)
training['CANDIDEMIA'] = y_training
training.head()

#Merge test set features and outcome in test set:
test = X_test.drop(X_test.columns[[0,3,4,5]], axis=1)
test['CANDIDEMIA'] = y_test
test.head()

#Definition of the imputer:
imputer = IterativeImputer(KNeighborsRegressor(n_neighbors=5),
                           sample_posterior=False, 
                           max_iter=100,
                           tol=0.05,
                           n_nearest_features=None,
                           initial_strategy='most_frequent',
                           imputation_order='random',
                           random_state=100)

#Fit of the imputer on training set and imputation:
training_imputed = imputer.fit_transform(training) 
training_imputed = pd.DataFrame(training_imputed)

print('Are missing values left in the training set?',training_imputed.isnull().values.any())

#Work on imputed training set:
training_imputed.columns = training.columns
training_imputed['30gg'] = training_imputed['30gg'] > 0.5
training_imputed['30gg'] = training_imputed['30gg']*1.
training_imputed.head()

#Imputation of the test set:
test_imputed = imputer.transform(test)
test_imputed = pd.DataFrame(test_imputed)

print('Are missing values left in the test set?',test_imputed.isnull().values.any())

#Work on imputed test set:
test_imputed.columns = test.columns
test_imputed['30gg'] = test_imputed['30gg'] > 0.5
test_imputed['30gg'] = test_imputed['30gg']*1.
test_imputed.head()

#### <font color='indianred'><b>EXPORT DATA SETS</b></font>

#Export data sets to csv file:
training_imputed.to_csv('04_Training_imputed.csv',sep=',') 
test_imputed.to_csv('04_Test_imputed.csv',sep=',')
