==> 05_Standardization.ipynb <==
# <font color='Crimson'><b>STANDARDIZATION</b></font>

#Import packages:
import pandas as pd
from sklearn.preprocessing import StandardScaler
import numpy as np

#Import imputed training set:
training_imputed = pd.read_csv('04_Training_imputed.csv',sep=',',index_col=0) 
training_imputed.head()

#Import imputed test set:
test_imputed = pd.read_csv('04_Test_imputed.csv',sep=',',index_col=0)
test_imputed.head()

#Change the order of the columns of the imputed training set:
training_imputed = training_imputed.reindex(columns=['CANDIDEMIA','MISTA','SESSO','30gg','ETA','BASOFILI','EOSINOFILI','LINFOCITI','MONOCITI',
                                       'NEUTROFILI','EMATOCRITO','EMOGLOBINA','GLOBULI_B','GLOBULI_R','PIASTRINE','APTT',
                                       'INR','TEMPO_PROTROMB','ACIDO_URICO','ALP','ALT','AST','BILIRUBINA_TOT','CREATININA','GGT',
                                       'LAD','UREA','GLUCOSIO','ALBUMINA','PCR','PROTEINE_TOT'])
training_imputed.head()

#Change the order of the columns of the imputed test set:
test_imputed = test_imputed.reindex(columns=['CANDIDEMIA','MISTA','SESSO','30gg','ETA','BASOFILI','EOSINOFILI','LINFOCITI','MONOCITI',
                                     'NEUTROFILI','EMATOCRITO','EMOGLOBINA','GLOBULI_B','GLOBULI_R','PIASTRINE','APTT',
                                     'INR','TEMPO_PROTROMB','ACIDO_URICO','ALP','ALT','AST','BILIRUBINA_TOT','CREATININA','GGT',
                                     'LAD','UREA','GLUCOSIO','ALBUMINA','PCR','PROTEINE_TOT'])
test_imputed.head()

#Definition of the scaler:
scaler = StandardScaler()

#Fitting the scaler on the continuous features of the training set and scale:
X_training_std = pd.DataFrame(scaler.fit_transform(training_imputed.iloc[:,4:training_imputed.shape[1]]))

#Scaling the continuous features of the test set:
X_test_std = pd.DataFrame(scaler.transform(test_imputed.iloc[:,4:test_imputed.shape[1]]))

#Add categorical features to the one just scaled (both sets) and change column names:
training_std = pd.DataFrame(np.concatenate((training_imputed.iloc[:,0:4], X_training_std), axis=1))
training_std.columns = training_imputed.columns
print(training_std.head())
print(test_std.head())
test_std = pd.DataFrame(np.concatenate((test_imputed.iloc[:,0:4], X_test_std), axis=1))
test_std.columns = test_imputed.columns

#### <font color='indianred'><b>EXPORT DATA SETS</b></font>

#Export data sets to csv file:
training_std.to_csv('05_Training_standard.csv',sep=',') 
test_std.to_csv('05_Test_standard.csv',sep=',')
